requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
spacy==3.7.2
scikit-learn==1.3.2
geopy==2.4.1
numpy>=1.24.0
sentence-transformers==2.2.2
transformers==4.36.2
huggingface-hub==0.19.4